+++ packages/algob/src/types.ts
@@ -1,6 +1,6 @@
 import { types as rtypes } from "@algo-builder/runtime";
 import { types as wtypes } from "@algo-builder/web";
-import algosdk, { LogicSig, modelsv2 } from "algosdk";
+import algosdk, { LogicSigAccount, modelsv2 } from "algosdk";
 
 import * as types from "./internal/core/params/argument-types";
 // Begin config types
@@ -311,7 +311,7 @@ export type AccountAddress = string;
 export interface LsigInfo {
   creator: AccountAddress
   contractAddress: string
-  lsig: LogicSig
+  lsig: LogicSigAccount
 }
 
 /**
@@ -553,7 +553,7 @@ export interface Deployer {
   /**
    * Creates an opt-in transaction for given ASA name, which must be defined in
    * `/assets/asa.yaml` file. The opt-in transaction is signed by the logic signature */
-  optInLsigToASA: (asa: string, lsig: LogicSig, flags: wtypes.TxParams) => Promise<void>
+  optInLsigToASA: (asa: string, lsig: LogicSigAccount, flags: wtypes.TxParams) => Promise<void>
 
   /**
    * Opt-In to stateful smart contract (SSC) for a single account
@@ -579,7 +579,7 @@ export interface Deployer {
    */
   optInLsigToApp: (
     appID: number,
-    lsig: LogicSig,
+    lsig: LogicSigAccount,
     payFlags: wtypes.TxParams,
     flags: rtypes.AppOptionalFlags) => Promise<void>
 
@@ -605,7 +605,7 @@ export interface Deployer {
    * @scTmplParams  Smart contract template parameters
    *     (used only when compiling PyTEAL to TEAL)
    */
-  loadLogic: (name: string, scTmplParams?: SCParams) => Promise<LogicSig>
+  loadLogic: (name: string, scTmplParams?: SCParams) => Promise<LogicSigAccount>
 
   /**
    * Returns ASCCache (with compiled code)
@@ -663,6 +663,7 @@ export interface AnyMap {
 export type PromiseAny = Promise<any>; // eslint-disable-line @typescript-eslint/no-explicit-any
 
 //  LocalWords:  configFile
+export type LogicSig = LogicSigAccount['lsig'];
 
 export interface DebuggerContext {
   tealFile?: string