+++ packages/algob/test/lib/dryrun.ts
@@ -146,7 +146,8 @@ describe("Debugging TEAL code using tealdbg", () => {
     try {
       await tealDebugger.run({ mode: ExecutionMode.APPLICATION, groupIndex: 5 });
     } catch (error) {
-      assert.equal(error.message, 'groupIndex(= 5) exceeds transaction group length(= 2)');
+      if (error instanceof Error) { assert.equal(error.message, 'groupIndex(= 5) exceeds transaction group length(= 2)'); }
+      console.error("An unexpected error occurred:", error);
     }
   });