+++ packages/runtime/src/index.ts
@@ -5,7 +5,7 @@ import { Interpreter } from "./interpreter/interpreter";
 import { loadASAFile, overrideASADef, parseASADef, validateASADefs, validateOptInAccNames } from "./lib/asa";
 import { getPathFromDirRecursive, loadFromYamlFileSilent, loadFromYamlFileSilentWithMessage, lsTreeWalk } from "./lib/files";
 import { checkIfAssetDeletionTx } from "./lib/txn";
-import { LogicSig } from "./logicsig";
+import { LogicSigAccount } from "./logicsig";
 import { parser } from "./parser/parser";
 import { Runtime } from "./runtime";
 import * as types from "./types";
@@ -17,7 +17,7 @@ export {
   validateOptInAccNames,
   Runtime,
   AccountStore,
-  LogicSig,
+  LogicSigAccount,
   checkIfAssetDeletionTx,
   loadFromYamlFileSilent,
   loadFromYamlFileSilentWithMessage,