+++ packages/web/src/types.ts
@@ -1,4 +1,4 @@
-import { Account as AccountSDK, LogicSig } from 'algosdk';
+import { Account as AccountSDK, LogicSigAccount } from 'algosdk';
 import * as z from 'zod';
 
 import type { ASADefSchema, ASADefsSchema } from "./types-input";
@@ -136,7 +136,7 @@ interface SignWithLsig {
   sign: SignType.LogicSignature
   fromAccount?: AccountSDK
   fromAccountAddr: AccountAddress
-  lsig: LogicSig
+  lsig: LogicSigAccount
   /** stateless smart contract args */
   args?: Uint8Array[]
 }
@@ -243,4 +243,23 @@ export interface RequestError extends Error {
     error?: Error
   }
   error?: Error
+}
+
+export interface FileError extends Error {
+  errno: number
+}
+
+// This function is used to check if given objects implements `FileError` interface
+export function isFileError (object: unknown): object is FileError {
+  return Object.prototype.hasOwnProperty.call(object, "errno");
+}
+
+// This function is used to check if given objects implements `RequestError` interface
+// https://www.technicalfeeder.com/2021/02/how-to-check-if-a-object-implements-an-interface-in-typescript/
+export function isRequestError (object: unknown): object is RequestError {
+  const res = Object.prototype.hasOwnProperty.call(object, "response.statusCode") &&
+   Object.prototype.hasOwnProperty.call(object, "response.text") &&
+   Object.prototype.hasOwnProperty.call(object, "response.body.message") &&
+   Object.prototype.hasOwnProperty.call(object, "response.error");
+  return res && Object.prototype.hasOwnProperty.call(object, "error");
 }