+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 ## Unreleased
 
 ### Improvements
++ Added shared space between contracts
++ Added tealv4 opcodes (`gload` and `gloads`) 
++ Added Tealv4 opcodes (`callsub` and `retsub`)
++ Added loop support in runtime
 * TEALv4 support in `@algo-builder/runtime`:
     * Added shared space between contracts (opcodes `gload` and `gloads`)
     * Dynamic Opcode Cost Evaluation
@@ -11,8 +15,13 @@
         b) User can pass id/offset for app/asset in for array references. For `tx.Accounts` you can pass address directly in teal code.
     * Byteslice arithmetic ops (`b+`, `b-`, `b*`, `b/`, `b%`, `b<`, `b>`, `b<=`, `b>=`, `b==`, `b!=`, `b\`, `b&`, `b^`, `b~`, `bzero`).
 
+
 ### Bug Fixes
 * Fixed `yarn add @algo-builder/web` (was failing because of missing dependency `zod` in packages/web).
+* Fix metadatahash type
+
+### API Breaking
+* Migrate from `LogicSig` to `LogicSigAccount`(Note: Loading lsig from file uses `LogicSig`, because `goal` stores it in LogicSig type format)
 
 ### API breaking
 * Rename `CallNoOpSSC` to `CallApp`.